+++ Users/cryptwallet/app/containers/SingleWalletDeposite/index.js
@@ -12,6 +12,8 @@ import { FormattedMessage } from 'react-intl';
 import { createStructuredSelector } from 'reselect';
 import { compose } from 'redux';
 import CopyAction from '../../images/icon/action/copy-action.png';
+import { getUser } from "../../store/auth/auth.actions";
+import { resetWalletStart } from "../../store/wallet/wallet.actions";
 
 import { toast } from 'react-toastify';
 
@@ -32,6 +34,7 @@ import {
   Typography,
 } from '@material-ui/core';
 import Form from '../../components/uiStyle/Form';
+import ResetWallet from "../../components/ResetWallet";
 import makeSelectSingleWalletDeposite from './selectors';
 import reducer from './reducer';
 import saga from './saga';
@@ -42,9 +45,12 @@ import Images from '../../components/uiStyle/Images';
 import './style.scss';
 import { address } from 'ip';
 
+const Row = [];
 /* eslint-disable react/prefer-stateless-function */
 export class SingleWalletDeposite extends React.Component {
   state = {
+    rwModalOpen: false,
+    wallet_name: "",
     qr_code: '',
     address: '',
     showAddress: false,
@@ -125,8 +131,58 @@ export class SingleWalletDeposite extends React.Component {
     });
   };
 
+  rwHandleClickOpen = () => {
+    this.setState({ rwModalOpen: true });
+    toast.warn("When you reset, you can't use old one!");
+  };
+  rwModalCloseHandler = () => {
+    this.setState({
+      rwModalOpen: false,
+    });
+  };
+
+  rwChangeHandler = (e) => {
+    this.setState({
+      [e.target.name]: e.target.value,
+    });
+  };
+
+  rwSubmitHandler = (e) => {
+    e.preventDefault();
+
+    // const {connectedUser} = this.props;
+    // const {addNewWallet} = this.props;
+    console.log("-this.props-------->>", this.props, "<<<-----------");
+    console.log("--this.state------>>", this.state, "<<<-----------");
+    const {getUser} = this.props;
+    const newWallet = {
+      id: this.props.row.id,
+      wallet_name: this.state.wallet_name,
+      wallet_address: this.state.address,
+      user_id: this.props.row.walletHolder,
+      balance: this.props.row.balance
+    };
+    const {resetWallet} = this.props;
+    // console.log('--getUser------->>',getUser,'<<<-----------');
+    // console.log('--connectedUser------->>',connectedUser,'<<<-----------');
+
+    Row.unshift(newWallet);
+
+    console.log('---this is rwSubmitHandler->>>', newWallet, '<<<----');
+    if (this.state.wallet_name === "") {
+      toast.error("Please give a valid info!");
+    } else {
+      this.setState({
+        wallet_name: "",
+        rwModalOpen: false,
+      });
+      resetWallet(newWallet);
+      getUser();
+    }
+  };
+
   render() {
-    const { address, qr_code, showAddress } = this.state;
+    const { address, qr_code, showAddress, rwModalOpen, wallet_name } = this.state;
 
     const AddressTable = (
       <Grid className="addressTable">
@@ -199,6 +255,15 @@ export class SingleWalletDeposite extends React.Component {
               Optimize Wallet
             </Typography>
             </p>
+            <p>
+              <Typography
+                onClick={this.rwHandleClickOpen}
+                className='generateNewAdress'
+                component='p'
+              >
+                Reset Wallet
+              </Typography>
+            </p>
             <Grid className="addressList">
               <Button
                 onClick={this.showAddressHandler}
@@ -211,6 +276,13 @@ export class SingleWalletDeposite extends React.Component {
             </Grid>
           </Grid>
         </Grid>
+        <ResetWallet
+          wallet_name={wallet_name}
+          rwModalOpen={rwModalOpen}
+          rwModalCloseHandler={this.rwModalCloseHandler}
+          rwChangeHandler={this.rwChangeHandler}
+          rwSubmitHandler={this.rwSubmitHandler}
+        />
       </Grid>
     );
   }
@@ -224,11 +296,11 @@ const mapStateToProps = createStructuredSelector({
   singleWalletDeposite: makeSelectSingleWalletDeposite(),
 });
 
-function mapDispatchToProps(dispatch) {
-  return {
-    dispatch,
-  };
-}
+const mapDispatchToProps = (dispatch) => ({
+  getUser: () => dispatch(getUser()),
+  resetWallet: (payload) => dispatch(resetWalletStart(payload)),
+});
+
 
 const withConnect = connect(
   mapStateToProps,